requirements.txt
# Core Dependencies (always installed)
neo4j>=5.13.0
google-generativeai>=0.8.5
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
opencv-python>=4.8.0
networkx>=3.0
matplotlib>=3.6.0
pyvis>=0.3.2

# Optional: Vision features (pip install -e ".[vision]")
# google-cloud-vision>=3.4.0

# Optional: Advanced visualization (pip install -e ".[visualization]")
# graphviz>=0.20.0

# Optional: Storage backends (pip install -e ".[storage]")
# redis>=4.0.0
# pymongo>=4.0.0
# psycopg2-binary>=2.9.0

# Optional: NLP features (pip install -e ".[nlp]")
# regex>=2022.0.0
# nltk>=3.8.0

# Optional: Vector storage (pip install -e ".[vector]")
# chromadb>=0.4.0

# Optional: Development tools (pip install -e ".[dev]")
# jupyter>=1.0.0
# jupyterlab>=3.0.0
# ipython>=8.0.0
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-mock>=3.10.0
# black>=22.0.0
# isort>=5.10.0
# mypy>=1.0.0
# flake8>=4.0.0

# Optional: Documentation (pip install -e ".[docs]")
# sphinx>=4.0.0
# sphinx-rtd-theme>=1.0.0

# Install all optional dependencies: pip install -e ".[all]"
